Module rtrees::rbtree

Structs

RBTree
A left-leaning red–black (LLRB) Tree, optimized for safety, simplicity of implementation, and augmentation. This tree is mimicking the behavior of a 2-3 tree. Full desciption of the tree design and complexity analysis is available in the paper titled Left-leaning Red-Black Trees by Robert Sedgewick. If you just need a normal non-augmentable tree-baesed map check std::collections::BTreeMap instead.

TreeIterator

TreeRefIterator

Traits

Augment
Used to recalculate augmented data stored in each node. This trait is mainly meant to be only implemented for RBTree before using the tree.

Type Definitions

LeftRightDataTuple
Tuple of 3 elements used with RBTree::mut_me. The first element is mutable reference to the left subtree, the second element is a mutable reference to the right subtree and the third element is a mutable reference to the data stored in the current node.