Type Definition rtrees::rbtree::LeftRightDataTuple

type LeftRightDataTuple<'a, K, A, V> = (&'a mut RBTree<K, A, V>, &'a mut RBTree<K, A, V>, &'a mut V);

Tuple of 3 elements used with RBTree::mut_me. The first element is mutable reference to the left subtree, the second element is a mutable reference to the right subtree and the third element is a mutable reference to the data stored in the current node.